Core - DependencyChecker now supports AnyCPU
- Remove the .Net Core specific implementation, was easier to unify them than first though
- Split checking for dlls into different folders to support AnyCPU
